requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.2
djangorestframework==3.11.0
drf-nested-routers==0.91
gunicorn==20.0.4
mysqlclient==1.3.13
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
whitenoise==5.0.1
